Create Kinematics
Specifically, it checks for:
The kinematic algorithm handles one robot, and one robot only

The robot has an arm.
The arm has a base frame, an end-effector frame, and urdf description.  In the code you can find useful documentation
and examples on how to build a robot model and calculate angles to develop some machine learning models.