Add files via upload
Grouping the datasets and give them some mechanical hierarchy models, print the elements and extract the data
For aerodynamics and robotic bodies are some of the priorities of this paper. 

Functional approach is needed as well as automotive instructions to draw some practical examples that can apply to spybotics and  Pydy libraries to model the robotic kinematics and find new projects behind the scenes. My aim as a programmer is apply the
Mathematical models and some examples in higher order dimensions with Scikit in robots to expand the capabilities that the field of robotics in python present and give them more analysis and new characteristics in a programming level foundation with python.